Move actions from AuthService to AuthAction.
