+++ packages/client-core/src/user/functions/userPatched.ts
@@ -8,7 +8,8 @@ import { WorldState } from '@xrengine/engine/src/networking/interfaces/WorldStat
 import { dispatchAction, getState } from '@xrengine/hyperflux'
 
 import { NotificationService } from '../../common/services/NotificationService'
-import { accessAuthState, AuthAction } from '../services/AuthService'
+import { AuthAction } from '../services/AuthAction'
+import { accessAuthState } from '../services/AuthService'
 import { accessUserState, UserAction } from '../services/UserService'
 
 const logger = multiLogger.child({ component: 'client-core:userPatched' })
