+++ packages/client/src/route/public.tsx
@@ -19,7 +19,8 @@ import { ProjectServiceReceptor } from '@xrengine/client-core/src/common/service
 import { LoadingCircle } from '@xrengine/client-core/src/components/LoadingCircle'
 import { InviteService, InviteServiceReceptor } from '@xrengine/client-core/src/social/services/InviteService'
 import { LocationServiceReceptor } from '@xrengine/client-core/src/social/services/LocationService'
-import { AuthService, AuthServiceReceptor } from '@xrengine/client-core/src/user/services/AuthService'
+import { AuthServiceReceptor } from '@xrengine/client-core/src/user/services/AuthAction'
+import { AuthService } from '@xrengine/client-core/src/user/services/AuthService'
 import {
   LocalStateServiceReceptor,
   StoredLocalAction,
